rateSwipe is now a compact script!
Thanks to the plotComputationVariableAgainstParameter() new plot
tool of PGW, most of the results processing code have been unified
and removed frrom the script. Transitioned to the defaults system
with the settings: now the parameters are removed from the
dictionary with the defaults not in its constant definition, but
dynamically right before the grid is defined.

Overall, the modifications greatly reduce the error proneness of
the script.

	modified:   pbsGridWalker
	modified:   rateSwipe.py